Register commands earlier
Fixes #103

We won't now wait for documents portal to open. Since it is still problematic
when rust-analyzer (or other extensions), launches first.

Also, don't show the ManifestManager's status item until there is some manifest
in the manifestMap. This is to avoid it being shown even the extension is
activated despite having no valid FlatpakManifest.

Add a when guard to every command (except show-output-terminal & select-manifest)
to only show when a manifest is active. Commands like rebuild will still be shown
even the last manifest is still deleted before this commit.

Rename too flatpakManifestFound to flatpakHasActiveManifest

Also reorder the menus.commandPalette in the package.json to have the same order
as in contributions.commands.