fix: Only write to workspace configuration when settings need to change
The workspace configuration in `.vscode/settings.json` is dependent on the order of `config.update()` calls. This causes a diff in my project, which is only a change in the order of the settings. By only calling `update()` when a change is needed, VSCode will not touch the `settings.json` once the configuration is correct.